requirements.txt

-i https://mirrors.aliyun.com/pypi/simple/
--extra-index-url https://pypi.org/simple
--only-binary=wandb
hydra-core==1.3.2
numpy
wandb
easydict
transformers
opencv-python
einops
thop
robosuite==1.4.0
bddl
future
matplotlib
cloudpickle
gym
numexpr
h5py
torchsde
torchdiffeq
imageio[ffmpeg]
open3d
ftfy
termcolor